Rewrite without internal state
Unfortunately we have to give up on the idea that we can capture all
events that lead to state changes.

While the original concept worked on desktop computers, it gets really
difficult once you have to support touch devices and non-keyboard input
events triggered by IMEs (e.g. touch keyboards or japanese input
methods). The problem is that the events can not always be captured,
sometimes you just know that "something changed", without having details
on what exactly changed.

So let's leave the state management to the browser. Instead, provide a
method to refresh the caret position, methods to insert text and block
elements, and a method to extract the current text from the compose
area.