Improve selection handling
When calling `store_selection_range`, only clear the selection when
there is no range at all in the document. If there is a range, but it's
outside the document, then keep the previous selection.

This allows to switch the focus from the area to another object in the
page, followed by an insertion at the correct position.